Add Orders Helper Module
